+++ package.json
@@ -10,7 +10,8 @@
     "require": "./build/main.cjs"
   },
   "scripts": {
-    "test": "mocha test/*.test.js test/test.js",
+    "test": "mocha",
+    "test:all": "mocha test/*.test.js test/test.js",
     "build": "rollup -c rollup.cjs.config.js"
   },
   "keywords": [
@@ -32,6 +33,8 @@
     "eslint": "^6.8.0",
     "ffjavascript": "^0.2.2",
     "mocha": "^8.1.3",
+    "chai": "^4.3.6",
+    "chai-as-promised": "^7.1.1",
     "rollup": "^2.20.0",
     "rollup-plugin-commonjs": "^10.1.0",
     "rollup-plugin-node-resolve": "^5.2.0",
